refactor: rewrite barcode mask function
